some frontend job for admin
Took 3 hours 18 minutes

+++ src/main/java/com/archpj/getatestbot/database/QuizResultsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface QuizResultsRepository extends JpaRepository<QuizResult, Long> {
+public interface QuizResultsRepository extends JpaRepository<QuizResult, Integer> {
 
 }
